requirements.txt
Django==2.0.6
djangorestframework==3.9.2
djangorestframework-jwt==1.11.0
PyMySQL==0.8.1
drf-yasg==1.15.0
swagger-spec-validator==2.1.0
flex==6.13.2
httpie==0.9.9
PyJWT==1.6.4
gunicorn==19.8.1
pyotp==2.2.6
django-redis==4.9.0
django-filter==1.1.0
pylint==1.9.2
autopep8==1.3.5
msgpack==0.5.6
django-cors-headers==2.3.0
django-crontab==0.7.1
django-rest-swagger==2.2.0
django-softdelete-it==0.1
django-aliyun-oss2-storage==0.1.5
xlwt==1.3.0
redis==2.10.6
drf-renderer-xlsx==0.3.2
